fixed unintended removal of environment variables from user profile if one Bench environment is registered in the user profile and the others are not, then the others would clean the profile on environment update

+++ BenchManager/BenchLib/BenchEnvironment.cs
@@ -339,6 +339,13 @@ namespace Mastersign.Bench
         /// </remarks>
         public void UnregisterFromUserProfile()
         {
+            var currentEnvRootPath = Environment.GetEnvironmentVariable("BENCH_HOME");
+            if (!string.Equals(currentEnvRootPath, Config.BenchRootDir,
+                StringComparison.InvariantCultureIgnoreCase))
+            {
+                return;
+            }
+
             DeleteEnvironmentVar("BENCH_VERSION");
             DeleteEnvironmentVar("BENCH_HOME");
             DeleteEnvironmentVar("BENCH_APPS");
